some stuff

+++ mac-code/include/usr/robot.h
@@ -79,8 +79,8 @@ class Robot {
         class Actions {
             public:
                 static int timesSorted;
-                static void setIntake (int dir, Intake_Action type);
-                static void setIntakeFor (int dir, Intake_Action type, int time, bool async=true);
+                static void setIntake (float dir, Intake_Action type);
+                static void setIntakeFor (float dir, Intake_Action type, int time, bool async=true);
 
                 static void FlingRing(bool async, float for_=100, float speed=127);
